backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-pptx>=0.6.23
google-genai>=0.3.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
